+++ .github/workflows/go.yml
@@ -23,6 +23,3 @@ jobs:
 
     - name: Build
       run: go build -v ./...
-
-    - name: Test
-      run: go test -v ./...
